+++ src/interfaces/hub/core/IRewardTreasury.sol
@@ -2,5 +2,5 @@
 pragma solidity ^0.8.26;
 
 interface IRewardTreasury {
-  function deposit(address asset, uint256 amount, uint256 timestamp) external;
+  function deposit(address asset, uint256 amount, uint48 timestamp) external;
 }
